requirements.txt
openai==1.34.0
python-dotenv~=1.0.1
streamlit~=1.35.0
pymysql~=1.1.1